requirements.txt
boto3
pyyaml
paramiko
scp
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
moto>=4.0.0
